CRA really doesn't play well in monorepo, replacing it ...
with a custom toolchain setup which uses webpack to pipe the output of
`tsc` (using `ts-loader`) through to babel for final bundling and
output.

The reasoning behind this is, in dev mode to leverage TypeScript's
built in project references to keep packages separated, and only have
to run a single build command to build everything, having `tsc`
resolve dependencies and building those as well.

Production builds still need to be figured out, but I forsee them just
referencing the published packages on NPM