Use context from the *most recent* output sent to the console for the interactive view.
The `from_start` parameter really only makes sense for the non-streaming
GET endpoint, and I'm working on removing it from Nexus (omicron#3147)
and the CLI (oxide.rs#190) accordingly.

As it stands, if an instance lives long enough to output more than 2MiB
to its serial port, asking for the stream with `?from_start=0` will cause
it to unconditionally fail.