Show a message that this site requires JavaScript (#2989)
* Show a message that this site requires JavaScript

Adds a noscript tag that displays a message that this site requires
JavaScript to run.

The easiest way to test this locally is to use Chrome since it can
disable JavaScript per tab instead of globally. Open the dev console and
click the gear icon and check "Disable "JavaScript" in the Debugger
section.

I skipped adding a test since it is a static change and playwright
requires js to run.

resolves #2928

* fill in noscript message more and style it

* Turns out we do actually have the CSS in prod

* forgot the external-link class

---------

Co-authored-by: David Crespo <david-crespo@users.noreply.github.com>